// cmportal-frontend/src/types/TUserForAdmin.ts
export type TUserForAdmin = {
    appid: string;
    userid: string;
    firstName: string;
    lastName: string;
    emailAddress: string;
    phoneNr: string;
    isDisabled: string;
    companyCode: string;
    companyName: string;
    userToken: string;
    urlUserImgProfileByAppid:string;
    hasAvatar: string;
}